Convert Signup's defaultScope to a scope function
This is necessary to fix the date checking - previously it only compared
to the server startup time. However, this forces us to put
.scope('active') in almost _every_ use of Signup. We may want to
consider abandoning the scope here and just use the cron job to delete
expired signups.

The exact same bug also applies to Event, but will hopefully be dealt
with in #44.